Fix build Qt 6.2 MSVC2019 Windows

+++ src/service/service.h
@@ -168,11 +168,11 @@ private:
 
 //! Overload for qHash
 Q_DECL_CONST_FUNCTION Q_DECL_CONSTEXPR inline uint qHash(QtService::Service::CommandResult key, uint seed = 0) Q_DECL_NOTHROW {
-	return ::qHash(static_cast<int>(key), seed);
+    return static_cast<uint>(::qHash(static_cast<int>(key), seed));
 }
 //! Overload for qHash
 Q_DECL_CONST_FUNCTION Q_DECL_CONSTEXPR inline uint qHash(QtService::Service::TerminalMode key, uint seed = 0) Q_DECL_NOTHROW {
-	return ::qHash(static_cast<int>(key), seed);
+    return static_cast<uint>(::qHash(static_cast<int>(key), seed));
 }
 
 template<typename TFunction>
