+++ src/service/servicebackend.h
@@ -90,7 +90,7 @@ private:
 
 //! Overload for qHash
 Q_DECL_CONST_FUNCTION Q_DECL_CONSTEXPR inline uint qHash(QtService::ServiceBackend::ServiceCommand key, uint seed = 0) Q_DECL_NOTHROW {
-	return ::qHash(static_cast<int>(key), seed);
+    return static_cast<uint>(::qHash(static_cast<int>(key), seed));
 }
 
 template<typename TRet, typename... TArgs>
