+++ src/service/servicecontrol.h
@@ -201,15 +201,15 @@ private:
 
 //! Overload for qHash
 Q_DECL_CONST_FUNCTION Q_DECL_CONSTEXPR inline uint qHash(QtService::ServiceControl::SupportFlags key, uint seed = 0) Q_DECL_NOTHROW {
-	return ::qHash(static_cast<int>(key), seed);
+    return static_cast<uint>(::qHash(static_cast<int>(key), seed));
 }
 //! Overload for qHash
 Q_DECL_CONST_FUNCTION Q_DECL_CONSTEXPR inline uint qHash(QtService::ServiceControl::Status key, uint seed = 0) Q_DECL_NOTHROW {
-	return ::qHash(static_cast<int>(key), seed);
+    return static_cast<uint>(::qHash(static_cast<int>(key), seed));
 }
 //! Overload for qHash
 Q_DECL_CONST_FUNCTION Q_DECL_CONSTEXPR inline uint qHash(QtService::ServiceControl::BlockMode key, uint seed = 0) Q_DECL_NOTHROW {
-	return ::qHash(static_cast<int>(key), seed);
+    return static_cast<uint>(::qHash(static_cast<int>(key), seed));
 }
 
 // ------------- Generic Implementations -------------
